Addressed most of the comments.
