+++ core/templates/pages/release-coordinator-page/beam-jobs-tab/beam-jobs-tab.component.html
@@ -33,7 +33,7 @@
   </div>
 
   <div class="job-runs-table-container mat-elevation-z8">
-    <table class="job-runs-table" mat-table recycleRows [dataSource]="filteredBeamJobRuns">
+    <table class="job-runs-table e2e-test-beam-jobs-table" mat-table recycleRows [dataSource]="filteredBeamJobRuns">
 
       <!-- Status Column -->
       <ng-container matColumnDef="run_status">
@@ -41,7 +41,11 @@
         <td mat-cell *matCellDef="let jobRun">
           <mat-icon class="job-status-mat-icon"
                     [matTooltip]="jobRun.getJobStatusTooltipString()"
-                    [color]="jobRun.getJobStatusMaterialThemeColor()">
+                    [color]="jobRun.getJobStatusMaterialThemeColor()"
+                    [ngClass]="{
+                      'e2e-test-job-status-success': jobRun.getJobStatusMaterialIconCode() === 'check_circle',
+                      'e2e-test-job-status-failure': jobRun.getJobStatusMaterialIconCode() !== 'check_circle'
+                    }">
             {{ jobRun.getJobStatusMaterialIconCode() }}
           </mat-icon>
         </td>
